+++ python/python/tests/tests_server_modules/test_json.py
@@ -359,3 +359,130 @@ class TestJson:
 
         with pytest.raises(RequestError):
             await json.clear(glide_client, "non_existing_key", ".")
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_json_numincrby(self, glide_client: TGlideClient):
+        key = get_random_string(10)
+
+        json_value = {
+            "key1": 1,
+            "key2": 3.5,
+            "key3": {"nested_key": {"key1": [4, 5]}},
+            "key4": [1, 2, 3],
+            "key5": 0,
+            "key6": "hello",
+            "key7": None,
+            "key8": {"nested_key": {"key1": 69}},
+            "key9": 1.7976931348623157e308,
+        }
+
+        # Set the initial JSON document at the key
+        assert await json.set(glide_client, key, "$", OuterJson.dumps(json_value)) == OK
+
+        # Test JSONPath
+        # Increment integer value (key1) by 5
+        result = await json.numincrby(glide_client, key, "$.key1", 5)
+        assert result == b"[6]"  # Expect 1 + 5 = 6
+
+        # Increment float value (key2) by 2.5
+        result = await json.numincrby(glide_client, key, "$.key2", 2.5)
+        assert result == b"[6]"  # Expect 3.5 + 2.5 = 6
+
+        # Increment nested object (key3.nested_key.key1[0]) by 7
+        result = await json.numincrby(glide_client, key, "$.key3.nested_key.key1[1]", 7)
+        assert result == b"[12]"  # Expect 4 + 7 = 12
+
+        # Increment array element (key4[1]) by 1
+        result = await json.numincrby(glide_client, key, "$.key4[1]", 1)
+        assert result == b"[3]"  # Expect 2 + 1 = 3
+
+        # Increment zero value (key5) by 10.23 (float number)
+        result = await json.numincrby(glide_client, key, "$.key5", 10.23)
+        assert result == b"[10.23]"  # Expect 0 + 10.23 = 10.23
+
+        # Increment a string value (key6) by a number
+        result = await json.numincrby(glide_client, key, "$.key6", 99)
+        assert result == b"[null]"  # Expect null
+
+        # Increment a None value (key7) by a number
+        result = await json.numincrby(glide_client, key, "$.key7", 51)
+        assert result == b"[null]"  # Expect null
+
+        # Check increment for all numbers in the document using JSON Path (First Null: key3 as an entire object. Second Null: The path checks under key3, which is an object, for numeric values).
+        result = await json.numincrby(glide_client, key, "$..*", 5)
+        assert (
+            result
+            == b"[11,11,null,null,15.23,null,null,null,1.7976931348623157e+308,null,null,9,17,6,8,8,null,74]"
+        )
+
+        # Check for multiple path match in enhanced
+        result = await json.numincrby(glide_client, key, "$..key1", 1)
+        assert result == b"[12,null,75]"
+
+        # Check for non existent path in JSONPath
+        result = await json.numincrby(glide_client, key, "$.key10", 51)
+        assert result == b"[]"  # Expect Empty Array
+
+        # Check for non existent key in JSONPath
+        with pytest.raises(RequestError):
+            await json.numincrby(glide_client, "non_existent_key", "$.key10", 51)
+
+        # Check for Overflow in JSONPath
+        with pytest.raises(RequestError):
+            await json.numincrby(glide_client, key, "$.key9", 1.7976931348623157e308)
+
+        # Decrement integer value (key1) by 12
+        result = await json.numincrby(glide_client, key, "$.key1", -12)
+        assert result == b"[0]"  # Expect 12 - 12 = 0
+
+        # Decrement integer value (key1) by 0.5
+        result = await json.numincrby(glide_client, key, "$.key1", -0.5)
+        assert result == b"[-0.5]"  # Expect 0 - 0.5 = -0.5
+
+        # Test Legacy Path
+        # Increment float value (key1) by 5 (integer)
+        result = await json.numincrby(glide_client, key, "key1", 5)
+        assert result == b"4.5"  # Expect -0.5 + 5 = 4.5
+
+        # Decrement float value (key1) by 5.5 (integer)
+        result = await json.numincrby(glide_client, key, "key1", -5.5)
+        assert result == b"-1"  # Expect 4.5 - 5.5 = -1
+
+        # Increment int value (key2) by 2.5 (a float number)
+        result = await json.numincrby(glide_client, key, "key2", 2.5)
+        assert result == b"13.5"  # Expect 11 + 2.5 = 13.5
+
+        # Increment nested value (key3.nested_key.key1[0]) by 7
+        result = await json.numincrby(glide_client, key, "key3.nested_key.key1[0]", 7)
+        assert result == b"16"  # Expect 9 + 7 = 16
+
+        # Increment array element (key4[1]) by 1
+        result = await json.numincrby(glide_client, key, "key4[1]", 1)
+        assert result == b"9"  # Expect 8 + 1 = 9
+
+        # Increment a float value (key5) by 10.2 (a float number)
+        result = await json.numincrby(glide_client, key, "key5", 10.2)
+        assert result == b"25.43"  # Expect 15.23 + 10.2 = 25.43
+
+        # Check for multiple path match in legacy and assure that the result of the last updated value is returned
+        result = await json.numincrby(glide_client, key, "..key1", 1)
+        assert result == b"76"
+
+        # Check if the rest of the key1 path matches were updated and not only the last value
+        result = await json.get(glide_client, key, "$..key1")
+        assert (
+            result == b"[0,[16,17],76]"
+        )  # First is 0 as 0 + 0 = 0, Second doesn't change as its an array type (non-numeric), third is 76 as 0 + 76 = 0
+
+        # Check for non existent path in legacy
+        with pytest.raises(RequestError):
+            await json.numincrby(glide_client, key, ".key10", 51)
+
+        # Check for non existent key in legacy
+        with pytest.raises(RequestError):
+            await json.numincrby(glide_client, "non_existent_key", ".key10", 51)
+
+        # Check for Overflow in legacy
+        with pytest.raises(RequestError):
+            await json.numincrby(glide_client, key, ".key9", 1.7976931348623157e308)